LayoutTests/platform/wk2/http/tests/media/video-auth-with-allowCrossOriginSubresourcesToAskForCredentials-expected.txt

127.0.0.1:8000 - didReceiveAuthenticationChallenge - ProtectionSpaceAuthenticationSchemeHTTPBasic - Responding with username:password
localhost:8000 - didReceiveAuthenticationChallenge - ProtectionSpaceAuthenticationSchemeHTTPBasic - Responding with username:password

Tests that the media player sends authorization credentials when requesting a media file.
Testing same domain (127.0.0.1)
EVENT(canplay)
Testing cross domain (localhost)
EVENT(canplay)
END OF TEST
